# Coelsa/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Simulador COELSA — Cheques y Echeqs
  version: 1.0.0
  description: |
    Microservicio que simula el rol de COELSA (cámara electrónica de compensación)
    limitado a la gestión de cheques físicos (identificados por CMC7) y echeqs
    (identificados por IDECHEQ).

    **Convenciones clave:**
    - La creación es idempotente: enviar el header `Idempotency-Key` (GUID).
    - Las consultas por CUIT/CUIL están separadas por tipo y son paginadas.
    - Los errores se responden con ProblemDetails (RFC 7807) en español.
servers:
  - url: http://localhost:8080
    description: docker compose (local)
...
paths:
...
  /api/v1/echeqs:
    post:
      tags: [Echeqs]
      operationId: crearEcheq
      summary: Crea un echeq (idempotente); el simulador asigna IDECHEQ y CMC7
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearEcheqRequest'
      responses:
        '201':
          description: Echeq creado (incluye IDECHEQ y CMC7 asignados).
          headers:
            Idempotent-Replay:
              $ref: '#/components/headers/IdempotentReplay'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EcheqResponse'
        '200':
          description: Replay idempotente (misma key + mismo body).
          headers:
            Idempotent-Replay:
              $ref: '#/components/headers/IdempotentReplay'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EcheqResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
    get:
      tags: [Echeqs]
      operationId: listarEcheqs
      summary: Lista echeqs por CUIT/CUIL (librador o beneficiario) con filtros opcionales, paginado
      parameters:
        - $ref: '#/components/parameters/Cuit'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: cbu
          in: query
          required: false
          description: Filtra por CBU de la cuenta de débito.
          schema:
            type: string
            pattern: '^\d{22}$'
        - name: estado
          in: query
          required: false
          description: Filtra por estado.
          schema:
            $ref: '#/components/schemas/EstadoInstrumento'
        - name: desdeEmision
          in: query
          required: false
          description: Emisión desde (el rango por eje no puede superar 360 días).
          schema:
            type: string
            format: date
        - name: hastaEmision
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: desdeVencimiento
          in: query
          required: false
          description: Vencimiento (fecha de pago) desde.
          schema:
            type: string
            format: date
        - name: hastaVencimiento
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: numeroCheque
          in: query
          required: false
          description: Número de cheque de la chequera.
          schema:
            type: integer
            format: int32
            minimum: 1
      responses:
        '200':
          description: Página de echeqs.
          content:
            application/json:
          schema:
            $ref: '#/components/schemas/PagedEcheqResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /api/v1/echeqs/{idecheq}:
    get:
      tags: [Echeqs]
      operationId: obtenerEcheq
      summary: Obtiene un echeq por su IDECHEQ
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      responses:
        '200':
          description: Echeq encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EcheqResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Echeqs]
      operationId: eliminarEcheq
      summary: Baja lógica de un echeq por IDECHEQ
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      responses:
        '204':
          description: Baja lógica aplicada.
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/echeqs/{idecheq}/estado:
    patch:
      tags: [Echeqs]
      operationId: cambiarEstadoEcheq
      summary: Cambia el estado de un echeq
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambiarEstadoRequest'
      responses:
        '200':
          description: Estado actualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EcheqResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/aceptacion:
    post:
      tags: [Echeqs]
      operationId: aceptarEcheq
      summary: Aceptación o repudio del beneficiario sobre un echeq pendiente
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AceptarEcheqRequest'
      responses:
        '200':
          description: Aceptado (Emitido) o repudiado (con motivo guardado).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EcheqResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/certificado:
    get:
      tags: [Echeqs]
      operationId: obtenerCertificado
      summary: Certificado para ejercer acciones civiles (solo echeqs rechazados)
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      responses:
        '200':
          description: CUD determinista + datos del rechazo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertificadoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/endosos:
    get:
      tags: [Echeqs]
      operationId: listarEndosos
      summary: Cadena de endosos del echeq para trazabilidad
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      responses:
        '200':
          description: Cadena ordenada por orden.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EndosoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Echeqs]
      operationId: proponerEndoso
      summary: Propone un endoso nominativo a un CUIT (pendiente de admisión)
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProponerEndosoRequest'
      responses:
        '201':
          description: Endoso propuesto.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndosoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/endosos/{orden}:
    get:
      tags: [Echeqs]
      operationId: obtenerEndoso
      summary: Obtiene un endoso por su orden en la cadena
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
        - $ref: '#/components/parameters/OrdenEndoso'
      responses:
        '200':
          description: Endoso encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndosoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Echeqs]
      operationId: anularEndoso
      summary: Anulación de un endoso propuesto por el endosante
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
        - $ref: '#/components/parameters/OrdenEndoso'
      responses:
        '204':
          description: Endoso anulado.
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/endosos/{orden}/admision:
    post:
      tags: [Echeqs]
      operationId: resolverEndoso
      summary: Admisión o repudio de un endoso propuesto (solo el endosatario)
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
        - $ref: '#/components/parameters/OrdenEndoso'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolverEndosoRequest'
      responses:
        '200':
          description: Endoso admitido o repudiado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndosoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/devoluciones:
    get:
      tags: [Echeqs]
      operationId: listarDevoluciones
      summary: Pedidos de devolución del echeq para trazabilidad
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      responses:
        '200':
          description: Pedidos ordenados por número.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DevolucionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Echeqs]
      operationId: solicitarDevolucion
      summary: Solicita la devolución del echeq (cualquier integrante de la cadena)
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolicitarDevolucionRequest'
      responses:
        '201':
          description: Pedido solicitado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevolucionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/devoluciones/{numero}:
    get:
      tags: [Echeqs]
      operationId: obtenerDevolucion
      summary: Obtiene un pedido de devolución por su número
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
        - $ref: '#/components/parameters/NumeroDevolucion'
      responses:
        '200':
          description: Pedido encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevolucionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Echeqs]
      operationId: anularDevolucion
      summary: Anulación de un pedido solicitado por el solicitante
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
        - $ref: '#/components/parameters/NumeroDevolucion'
      responses:
        '204':
          description: Pedido anulado.
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/devoluciones/{numero}/resolucion:
    post:
      tags: [Echeqs]
      operationId: resolverDevolucion
      summary: Aceptación o rechazo del pedido por el tenedor actual
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
        - $ref: '#/components/parameters/NumeroDevolucion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolverDevolucionRequest'
      responses:
        '200':
          description: Pedido aceptado o rechazado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DevolucionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/cesiones:
    get:
      tags: [Echeqs]
      operationId: listarCesiones
      summary: Cesiones del echeq "no a la orden" para trazabilidad
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      responses:
        '200':
          description: Cesiones ordenadas por número.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CesionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Echeqs]
      operationId: solicitarCesion
      summary: Solicita la cesión del echeq a un tercero (solo "no a la orden", en Emitido)
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolicitarCesionRequest'
      responses:
        '201':
          description: Cesión solicitada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CesionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/cesiones/{numero}:
    get:
      tags: [Echeqs]
      operationId: obtenerCesion
      summary: Obtiene una cesión por su número
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
        - $ref: '#/components/parameters/NumeroCesion'
      responses:
        '200':
          description: Cesión encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CesionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Echeqs]
      operationId: anularCesion
      summary: Anulación de una cesión solicitada por el cedente
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
        - $ref: '#/components/parameters/NumeroCesion'
      responses:
        '204':
          description: Cesión anulada.
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /api/v1/echeqs/{idecheq}/cesiones/{numero}/resolucion:
    post:
      tags: [Echeqs]
      operationId: resolverCesion
      summary: Aceptación o rechazo de la cesión por el cesionario
      parameters:
        - $ref: '#/components/parameters/IdEcheq'
        - $ref: '#/components/parameters/NumeroCesion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolverCesionRequest'
      responses:
        '200':
          description: Cesión aceptada o rechazada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CesionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: GUID único por operación de creación. Mismo key + mismo body = replay seguro.
      schema:
        type: string
        format: uuid
      example: 3f8a1c2e-9b4d-4f6a-8e2b-1d7c5a9e0f33
...
    Cuit:
      name: cuit
      in: query
      required: true
      description: CUIT/CUIL (librador o beneficiario). 11 dígitos con verificador módulo 11.
      schema:
        type: string
        pattern: '^\d{11}$'
      example: '20123456789'
...
  headers:
    IdempotentReplay:
      description: "'true' si la respuesta es un replay idempotente de una creación previa."
      schema:
        type: boolean
        default: false
  responses:
    BadRequest:
      description: Error de validación (body inválido, CUIT inválido, falta Idempotency-Key, etc.).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Conflict:
      description: Conflicto — CBU/CMC7/IDECHEQ duplicado con key distinta, misma Idempotency-Key con body distinto, chequera sin números disponibles, tope de endosos/cesiones o pedido/cesión pendiente.
...
    TooManyRequests:
      description: Se excedió el límite de solicitudes (rate limiting, RF-08). Reintentar cuando indique Retry-After.
      headers:
        Retry-After:
          description: Segundos hasta que se pueda reintentar.
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
  schemas:
...
    CrearEcheqRequest:
      type: object
      required: [cbuEmisor, caracter, tipoDocBeneficiario, nombreLibrador, nombreBeneficiario, cuitLibrador, cuitBeneficiario, monto, moneda, fechaEmision, fechaVencimiento]
      properties:
        cbuEmisor:
          type: string
          pattern: '^\d{22}$'
          description: CBU de la cuenta de débito (con chequera vigente). El librador debe ser su titular y la moneda debe coincidir.
          example: '0110001300000000000017'
        caracter:
          type: string
          enum: [AlaOrden, NoAlaOrden]
          description: Solo los 'A la orden' se endosan; los 'No a la orden' se transmiten por cesión.
          example: AlaOrden
        tipoDocBeneficiario:
          type: string
          enum: [CUIT, CUIL, CDI]
          description: Tipo de documento del beneficiario (11 dígitos con verificador).
          example: CUIT
        nombreLibrador:
          type: string
          maxLength: 120
          example: 'Alfa S.R.L.'
        nombreBeneficiario:
          type: string
          maxLength: 120
          example: 'Beta S.A.'
        concepto:
          type: string
          maxLength: 60
          nullable: true
          example: 'Pago a proveedores'
        motivo:
          type: string
          maxLength: 280
          nullable: true
        referencia:
          type: string
          maxLength: 60
          nullable: true
          example: 'FAC-123'
        emailNotificacion:
          type: string
          maxLength: 160
          nullable: true
          description: Email de aviso al beneficiario.
          example: 'cobros@demo.local'
        cuitLibrador:
          $ref: '#/components/schemas/Cuit'
        cuitBeneficiario:
          $ref: '#/components/schemas/Cuit'
        monto:
          $ref: '#/components/schemas/Monto'
        moneda:
          $ref: '#/components/schemas/Moneda'
        fechaEmision:
          type: string
          format: date
          example: '2026-09-05'
        fechaDiferimiento:
          type: string
          format: date
          nullable: true
          example: null
        fechaVencimiento:
          type: string
          format: date
          description: Obligatoria. Debe ser posterior a fechaEmision (y a fechaDiferimiento si viene).
          example: '2026-11-05'
...
    EcheqResponse:
      type: object
      properties:
        identificador:
          type: string
          pattern: '^[A-Z]{11}$'
          description: El IDECHEQ (11 letras mayúsculas) asignado por el simulador.
          example: ABCDEFGHIJK
        tipo:
          type: string
          enum: [Echeq]
        cbuEmisor:
          type: string
          pattern: '^\d{22}$'
          description: CBU de la cuenta de débito.
          example: '0110001300000000000017'
        numeroChequera:
          type: integer
          format: int32
          description: Número de la e-chequera usada (secuencial por cuenta).
          example: 1
        numeroCheque:
          type: integer
          format: int32
          description: Número de cheque reservado de la chequera (1-50).
          example: 7
        caracter:
          type: string
          enum: [AlaOrden, NoAlaOrden]
          description: Solo los 'A la orden' se endosan.
          example: AlaOrden
        modo:
          type: string
          enum: [Cruzado]
          description: Siempre cruzado (solo depósito en cuenta).
          example: Cruzado
        tipoDocBeneficiario:
          type: string
          enum: [CUIT, CUIL, CDI]
          example: CUIT
        nombreLibrador:
          type: string
          example: 'Alfa S.R.L.'
        nombreBeneficiario:
          type: string
          example: 'Beta S.A.'
        concepto:
          type: string
          nullable: true
        motivo:
          type: string
          nullable: true
        referencia:
          type: string
          nullable: true
        emailNotificacion:
          type: string
          nullable: true
        cmc7:
          type: string
          pattern: '^\d{30}$'
          description: CMC7 completo del echeq, generado por el simulador. Único.
          example: '011000114250000123400001234567'
        desgloseCmc7:
          type: object
          description: Tramos derivados del CMC7 (no se persisten).
          properties:
            banco:
              type: string
              example: '011'
            sucursal:
              type: string
              example: '0001'
            codigoPostal:
              type: string
              example: '1425'
            numeroCheque:
              type: string
              example: '00001234'
            numeroCuenta:
              type: string
              example: '00001234567'
        cuitLibrador:
          $ref: '#/components/schemas/Cuit'
        cuitBeneficiario:
          $ref: '#/components/schemas/Cuit'
        monto:
          $ref: '#/components/schemas/Monto'
        moneda:
          $ref: '#/components/schemas/Moneda'
        fechaEmision:
          type: string
          format: date
        fechaDiferimiento:
          type: string
          format: date
          nullable: true
        fechaVencimiento:
          type: string
          format: date
        estado:
          $ref: '#/components/schemas/EstadoInstrumento'
        motivoRechazo:
          allOf:
            - $ref: '#/components/schemas/MotivoRechazo'
          nullable: true
        motivoRepudio:
          type: string
          nullable: true
          description: Motivo informado por el beneficiario al repudiar.
        cantidadEndosos:
          type: integer
          format: int32
          example: 0
        fechaCreacion:
          type: string
          format: date-time
...
    Monto:
      type: number
      format: double
      exclusiveMinimum: 0
      example: 1500000.50
...